+++ experimental/packages/exporter-logs-otlp-proto/test/logHelper.ts
@@ -45,6 +45,7 @@ export const mockedReadableLogRecord: ReadableLogRecord = {
   attributes: {
     'some-attribute': 'some attribute value',
   },
+  droppedAttributesCount: 0,
   severityNumber: SeverityNumber.ERROR,
   severityText: 'error',
   body: 'some_log_body',